move colorscale from js_plotting_utils to plotting._utils
